Refactor actions and have separate actions for each root reducer.
