+++ src/actions/assetFilesActions.js
@@ -15,6 +15,7 @@ export const fetchFileConfigFailedAction = errorMessage => ({
 export const loadAppDataAction = () => ({
   type: 'LOAD_APP_DATA_ACTION',
 });
+
 export const loadBusCoordinatorsDataAction = () => ({
   type: 'LOAD_BUS_COORDINATORS_DATA_ACTION',
 });
@@ -28,6 +29,7 @@ export const loadBusCoordinatorsDataFailedAction = errorMessage => ({
   type: 'LOAD_BUS_COORDINATORS_DATA_FAILED_ACTION',
   errorMessage,
 });
+
 export const loadedAppDataSuccessAction = modeVariable => ({
   type: 'LOADED_APP_DATA_SUCCESS_ACTION',
   modeVariable,
@@ -37,4 +39,3 @@ export const loadAppDataFailedAction = errorMessage => ({
   type: 'LOAD_APP_DATA_FAILED_ACTION',
   errorMessage,
 });
-
