+++ src/components/routeComponents/splashPage/SplashPage.jsx
@@ -19,10 +19,12 @@ import Typography from 'pepcus-core/lib/Typography';
 import {
   fetchMemberDataAction,
   setAdminCredentialsAction,
-  setAdminLoginStateAction,
-  setHashLinkForNewRegistrationAction,
   setMemberCredentialsAction,
 } from 'actions/memberRegistrationActions';
+import {
+  setAdminLoginStateAction,
+  setHashLinkForNewRegistrationAction,
+} from 'actions/allMembersDataActions';
 import {
   getAdminId,
   getAdminPassword,
